requirements.txt
appdirs==1.4.0
bcrypt==3.1.2
blinker==1.4
cffi==1.9.1
dj-database-url==0.4.2
Django==1.10.5
django-cors-headers==2.0.0
django-mongodbforms==0.3.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.5.3
gunicorn==19.6.0
mongoengine==0.11.0
packaging==16.8
pycparser==2.17
pymongo==3.4.0
pyparsing==2.1.10
six==1.10.0
